Add lat, lon, elevation, and type information to core list
This resulted in discovering some mixed types in the core
database source data. I decided to fix this in the python
processing script (i.e. as far upstream as possible). The
keys are also now sorted and so the reactors.json can be
easily compared using something like `jq . reactors.json`
to get a pretty print output.