Update solar-plots.jsx
Actually deleting the  "+ 1" in line 32 makes the short segment of an analemma loop appear when start and end month are the same. With the + 1 no portion of the analemma loop appears when selecting the same start and end month. I am deleting the "+ 1" again. However, no portion of the loop appears when selecting the start and end month 01 (January).